remove MasterSlaveRuleConfigurationBeanDefinition
